requirements.txt
browser-use>=0.1.40
langchain-google-genai>=2.0.0
playwright>=1.40.0
python-dotenv>=1.0.0
asyncio
pydantic>=2.0.0
# langchain-groq>=0.1.0
langchain-core>=0.3.0